+++ src/ui-api.ts
@@ -569,7 +569,7 @@ export class UiApi {
     } catch (error: any) {
       this.log.error(`${error.code} error connecting to ${this.baseUrl + apiPath}`)
       if (error.code === 'ERR_BAD_REQUEST' && error.status === 404 && apiPath === ApiPluginEndpoints.getIgnoredPluginList) {
-        this.log.debug(`Error: ${JSON.stringify(error, undefined, 2)}`)
+        this.log.debug(`Error: ${error instanceof Error ? error.message : String(error)}`)
         this.log.warn('This feature requires a newer version of Homebridge UI. Please update to the latest version.')
       }
       return []